requirements.txt
streamlit
pandas
openpyxl
sqlalchemy
psycopg2-binary
bcrypt